Added method for casting nested structs
Added some additional state to structs (type) and struct instances that
better allows them to do the necessary introspection required for being
able to figure out upcasting to nested structs. This is important
because it allows us to cast classes to super-classes automatically in
functions. TODO: Incorporate this into ast_function_assembler.